feat: add SSO support for github (#122)
This PR implement feature of checking if user has SSO enabled before
adding the user to github group.

The procedure of doing this is:
1. Use GraphQL to Get all SAML identities within the given org.
2. When adding users to github team, check if the given team requires
SSO and if this user has saml identity in the organization.

This feature is tested in my test org with mapping a same google group
to two different github team with one requires SSO and the other one
don't.

Team that requires SSO only added member with saml identity. While the
doesn't require SSO added all team member.