feat: Reduce work done for ManyToOne SyncAll (#164)
In ManyToOne, the unit of work is based on the target groups, not the
source groups. The previous implementation kicked off concurrent
processing based on each known source group, which would cause the same
target groups to be handled multiple times.

---

Also update to Go 1.25 to take advantage of the newly added
`waitgroup.Go()` in stdlib.

Signed-off-by: James Alseth <jalseth@google.com>